APP RUNNING Mar 25th

+++ Cruise-main (1)/Cruise-main/practice cms/cms/src/main/java/com/example/cms/controller/exceptions/QuestionNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.cms.controller.exceptions;
 
 public class QuestionNotFoundException extends RuntimeException{
-    public QuestionNotFoundException(Long id) {
+    public QuestionNotFoundException(int id) {
         super("Could not find question " + id);
     }
 }
